requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
motor
beanie